# backend/src/models/Bot.ts
export type BotStatus = 'DISABLED' | 'ENABLED' | 'PAUSED';

export interface Bot {
  id: string;
  name: string;
  description?: string;
  status: BotStatus;
  created: number;
}

export interface CreateBotInput {
  id?: string;
  name: string;
  description?: string;
  status?: BotStatus;
}

export interface UpdateBotInput {
  id?: string;
  name?: string;
  description?: string;
  status?: BotStatus;
}